+++ src/api/core/index.ts
@@ -1,43 +1,25 @@
-import axios, { AxiosInstance, AxiosRequestConfig, Method } from 'axios';
+import axios, { AxiosInstance } from 'axios';
 
-const BASE_URL = process.env.NEXT_PUBLIC_BASE_URL;
+import { insertAccessToken } from '@/api/core/helper';
 
-const baseRequest = axios.create({
-  baseURL: BASE_URL,
-});
+const setInterceptors = (instance: AxiosInstance) => {
+  instance.interceptors.request.use(
+    (config) => {
+      return insertAccessToken(config);
+    },
+    (error) => Promise.reject(error),
+  );
 
-baseRequest.interceptors.request.use(
-  (request) => request,
-  (error) => Promise.reject(error),
+  instance.interceptors.response.use((response) => response);
+
+  return instance;
+};
+
+const http = setInterceptors(
+  axios.create({
+    baseURL: process.env.NEXT_PUBLIC_BASE_URL,
+    withCredentials: true,
+  }),
 );
-
-const handleRequest = (config: AxiosRequestConfig) => {
-  const token = localStorage.getItem('accessToken');
-
-  return token
-    ? ({
-        ...config,
-        headers: {
-          ...config.headers,
-          AccessToken: `Bearer ${JSON.parse(token)}`,
-        },
-      } as AxiosRequestConfig)
-    : config;
-};
-
-const createApiMethod =
-  (axiosInstance: AxiosInstance, methodType: Method) => (config: AxiosRequestConfig) => {
-    return axiosInstance({ ...handleRequest(config), method: methodType });
-  };
-
-export { baseRequest };
-
-const http = {
-  get: createApiMethod(baseRequest, 'GET'),
-  post: createApiMethod(baseRequest, 'POST'),
-  put: createApiMethod(baseRequest, 'PUT'),
-  patch: createApiMethod(baseRequest, 'PATCH'),
-  delete: createApiMethod(baseRequest, 'DELETE'),
-};
 
 export default http;
